schedulator_backend/requirements.txt
arrow==0.14.7
asgiref==3.2.3
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==3.0.2
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.10.3
ics==0.6
idna==2.8
IMAPClient==2.1.0
psycopg2-binary==2.8.4
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.4
sqlparse==0.3.0
TatSu==4.4.0
urllib3==1.25.6
